backend/requirements.txt
Flask
PyJWT
flask-limiter
SQLAlchemy
cryptography
reportlab
qrcode
Pillow
requests
redis>=4.0.0

